Updating the local `/_grow/` handlers to work when the routes reset.
When the routes are reset then the handlers are also lost for the `/_grow/` pages are lost. This change consolidates the handlers and triggers the dev handler hook to recreate the built-in dev routes.

Fixes #855
